arm64: dts: qcom: msm8916: Place wcnss/venus memory next to mba
The modem (mpss) firmware size varies a lot from device to device.
It probably depends on the features that are enabled in the firmware,
so there are often differences of several megabytes. Unfortunately,
the current memory layout makes it really inconvenient to increase
the memory region for the modem.

mpss/wcnss/venus are placed contiguously in memory, which means that
changing the mpss size requires moving the wcnss/venus memory as well.
The size for wcnss/venus seems consistent on all devices though
so this is normally not necessary.

To avoid this, move wcnss/venus to the end of the space between
mpss and mba. The reserved-memory is still largely contiguous
but this allows growing mpss without moving wcnss/venus.

The only device where this does not help is longcheer-l8150 because
of its weird non-relocatable wcnss firmware. However, since it is
the only device with this problem it does not seem worth optimizing
for that case.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>