FROMLIST: thermal: qcom: tsens: Add data for MSM8909
The MSM8909 SoC has 5 thermal sensors in a TSENS v0.1 block similar to
MSM8916, except that the bit offsets in the qfprom were changed.
Also, some fixed correction factors are needed as workaround because the
factory calibration apparently was not reliable enough.

Add the defines and calibration function for MSM8909 in the existing
tsens-v0_1.c driver to make the thermal sensors work on MSM8909.
The changes are derived from the original msm-3.18 kernel [1] from
Qualcomm but cleaned up and adapted to the driver in mainline.

[1]: https://git.codelinaro.org/clo/la/kernel/msm-3.18/-/blob/LA.UM.7.7.c26-08600-8x09.0/drivers/thermal/msm-tsens.c

Signed-off-by: Stephan Gerhold <stephan.gerhold@kernkonzept.com>
Link: https://lore.kernel.org/r/20220627131415.2868938-4-stephan.gerhold@kernkonzept.com