arm64: dts: qcom: msm8939-samsung-a7: Add touch key (v2)
Add the CORERIVER TC360 touch key together with the necessary
fixed regulator for it.

Note that for some reason Samsung decided to connect this to GPIOs
where no hardware I2C bus is available, so we need to fall back
to software bit-banging using i2c-gpio.

v2: Add suffix "-state" to pinctrl.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>