arm64: dts: qcom: msm8939: Drop DSI power-domains again
This was likely added to (partially) satisfy the DT schema but should
be redundant: The same power domain is already specified for the parent
"mdss" node. The DSI nodes are supposed to use the VDDCX power domain
together with an OPP table, so that the voltage is adjusted when higher
DSI clock frequencies are used.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>