genpd: qcom: rpmpd: Add MSM8917, MSM8937 and QM215
MSM8917 uses the SMPA2 and LDOA3 regulators provided by the PM8937 PMIC
for the VDDCX and VDDMX power domains in voltage level mode,
respectively. MSM8937 also uses this PMIC and the same regulators.

QM215 is typically paired with a PM8916 PMIC and uses its SMPA1 and
LDOA2 regulators in voltage level mode.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>