WIP: media: camss: Support stats stream for VFE 4.1
The VFE is able to produce a stream with various measurements needed for
3A (auto white balance, auto exposure correction and autofocus) algorithms.

Add a separate meta capture device to the CAMSS pipeline which allows
capturing these statistics when using the PIX line. Implement the Bayer
Exposure stream, which can be useful for automatic white balance and
exposure correction.

TODO: Implement it for other VFE versions, or at least make sure it
doesn't crash. Also fix some buffer management issues.